# cli/test/e2e/expected_openapi_spec.yaml
openapi: 3.0.1
info:
  title: tyger.server
  version: '1.0'
paths:
  /v1/clusters/:
    get:
      tags:
        - tyger.server
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cluster'
  '/v1/clusters/{name}':
    get:
      tags:
        - tyger.server
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cluster'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /v1/buffers:
    post:
      tags:
        - tyger.server
      operationId: createBuffer
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Buffer'
  '/v1/buffers/{id}':
    get:
      tags:
        - tyger.server
      operationId: getBufferById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Buffer'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  '/v1/buffers/{id}/access':
    post:
      tags:
        - tyger.server
      operationId: getBufferAccessString
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: writeable
          in: query
          schema:
            type: boolean
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BufferAccess'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  '/v1/codespecs/{name}':
    put:
      tags:
        - tyger.server
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Codespec'
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Codespec'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
    get:
      tags:
        - tyger.server
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Codespec'
  '/v1/codespecs/{name}/versions/{version}':
    get:
      tags:
        - tyger.server
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Codespec'
  /v1/runs:
    post:
      tags:
        - tyger.server
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
    get:
      tags:
        - tyger.server
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            format: int32
        - name: since
          in: query
          schema:
            type: string
        - name: _ct
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunPage'
  '/v1/runs/{runId}':
    get:
      tags:
        - tyger.server
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  '/v1/runs/{runId}/logs':
    get:
      tags:
        - tyger.server
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
        - name: timestamps
          in: query
          schema:
            type: boolean
        - name: tailLines
          in: query
          schema:
            type: integer
            format: int32
        - name: since
          in: query
          schema:
            type: string
        - name: follow
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: string
  /v1/metadata:
    get:
      tags:
        - tyger.server
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Metadata'
components:
  schemas:
    Buffer:
      type: object
      properties:
        id:
          type: string
          nullable: true
      additionalProperties: false
    BufferAccess:
      type: object
      properties:
        uri:
          type: string
          format: uri
          nullable: true
      additionalProperties: false
    BufferParameters:
      type: object
      properties:
        inputs:
          type: array
          items:
            type: string
          nullable: true
        outputs:
          type: array
          items:
            type: string
          nullable: true
      additionalProperties: false
    Cluster:
      type: object
      properties:
        name:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        nodePools:
          type: array
          items:
            $ref: '#/components/schemas/NodePool'
          nullable: true
      additionalProperties: false
    Codespec:
      required:
        - image
      type: object
      properties:
        kind:
          $ref: '#/components/schemas/CodespecKind'
        image:
          type: string
        command:
          type: array
          items:
            type: string
          nullable: true
        args:
          type: array
          items:
            type: string
          nullable: true
        workingDir:
          type: string
          nullable: true
        env:
          type: object
          additionalProperties:
            type: string
          nullable: true
        resources:
          $ref: '#/components/schemas/CodespecResources'
        maxReplicas:
          type: integer
          format: int32
        buffers:
          $ref: '#/components/schemas/BufferParameters'
        name:
          type: string
          nullable: true
        version:
          type: integer
          format: int32
      additionalProperties: false
    CodespecKind:
      enum:
        - 0
        - 1
      type: integer
      format: int32
    CodespecResources:
      type: object
      properties:
        cpu:
          type: string
          nullable: true
        memory:
          type: string
          nullable: true
        gpu:
          type: string
          nullable: true
      additionalProperties: false
    ErrorBody:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorInfo'
      additionalProperties: false
    ErrorInfo:
      type: object
      properties:
        code:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
      additionalProperties: false
    Metadata:
      type: object
      properties:
        authority:
          type: string
          nullable: true
        audience:
          type: string
          nullable: true
      additionalProperties: false
    NodePool:
      type: object
      properties:
        name:
          type: string
          nullable: true
        vmSize:
          type: string
          nullable: true
      additionalProperties: false
    Run:
      required:
        - job
      type: object
      properties:
        job:
          $ref: '#/components/schemas/RunCodeTarget'
        worker:
          $ref: '#/components/schemas/RunCodeTarget'
        timeoutSeconds:
          type: integer
          format: int32
          nullable: true
        cluster:
          type: string
          nullable: true
        id:
          type: integer
          format: int64
        status:
          type: string
          nullable: true
        reason:
          type: string
          nullable: true
        runningCount:
          type: integer
          format: int32
          nullable: true
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
    RunCodeTarget:
      required:
        - codespec
      type: object
      properties:
        codespec:
          type: string
        buffers:
          type: object
          additionalProperties:
            type: string
          nullable: true
        nodePool:
          type: string
          nullable: true
        replicas:
          type: integer
          format: int32
      additionalProperties: false
    RunPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Run'
          nullable: true
        nextLink:
          type: string
          format: uri
          nullable: true
      additionalProperties: false